requirement.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
chromadb==0.5.5
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1